+++ itests/src/test/resources/testconfiguration.properties
@@ -222,7 +222,6 @@ mr.query.files=\
   mapjoin_subquery2.q,\
   mapjoin_test_outer.q,\
   masking_5.q,\
-  msck_repair_filter.q,\
   nonmr_fetch.q,\
   nonreserved_keywords_input37.q,\
   parenthesis_star_by.q,\
