+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/metastore/HiveMetaStoreChecker.java
@@ -375,17 +375,26 @@ public class HiveMetaStoreChecker {
     // now check the table folder and see if we find anything
     // that isn't in the metastore
     Set<Path> allPartDirs = new HashSet<Path>();
-    Set<Path> partDirs = new HashSet<Path>();
     List<FieldSchema> partColumns = table.getPartitionKeys();
     checkPartitionDirs(tablePath, allPartDirs, Collections.unmodifiableList(getPartColNames(table)));
 
     if (filterExp != null) {
       PartitionExpressionProxy expressionProxy = createExpressionProxy(conf);
+      if (expressionProxy instanceof MsckPartitionExpressionProxy) {
+        throw new MetastoreException("Unsupported expressionProxy type. " +
+            "This happens when remote metastore setup is used. Try with embedded metastore");
+      }
       List<String> paritions = new ArrayList<>();
+      Set<Path> partDirs = new HashSet<Path>();
+      String tablePathStr = tablePath.toString();
       for (Path path : allPartDirs) {
         // remove the table's path from the partition path
         // eg: <tablePath>/p1=1/p2=2/p3=3 ---> p1=1/p2=2/p3=3
-        paritions.add(path.toString().substring(tablePath.toString().length() + 1));
+        if (tablePathStr.endsWith("/")) {
+          paritions.add(path.toString().substring(tablePathStr.length()));
+        } else {
+          paritions.add(path.toString().substring(tablePathStr.length() + 1));
+        }
       }
       // Remove all partition paths which does not matches the filter expression.
       expressionProxy.filterPartitionsByExpr(partColumns, filterExp,
@@ -394,7 +403,7 @@ public class HiveMetaStoreChecker {
       // now the partition list will contain all the paths that matches the filter expression.
       // add them back to partDirs.
       for (String path : paritions) {
-        partDirs.add(new Path(tablePath.toString() + "/" + path));
+        partDirs.add(new Path(tablePath, path));
       }
       allPartDirs = partDirs;
     }
